Simplify Program.cs with extension methods
