+++ ClinicManagement/src/ClinicManagement.Api/Program.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
 using System.Reflection;
 using BlazorShared;
 using ClinicManagement.Api;
@@ -7,10 +6,8 @@ using ClinicManagement.Core.Aggregates;
 using ClinicManagement.Core.Interfaces;
 using ClinicManagement.Infrastructure;
 using ClinicManagement.Infrastructure.Data;
-using ClinicManagement.Infrastructure.Messaging;
 using FastEndpoints;
 using FastEndpoints.Swagger;
-using MassTransit;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.ResponseCompression;
@@ -18,7 +15,6 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -60,28 +56,7 @@ builder.Services.AddResponseCompression(opts =>
 builder.Services.AddAutoMapper(typeof(Program).Assembly);
 builder.Services.AddSwaggerGenCustom();
 
-// configure messaging
-var messagingConfig = builder.Configuration.GetSection("RabbitMq");
-builder.Services.Configure<RabbitMqConfiguration>(messagingConfig);
-builder.Services.AddScoped<IMessagePublisher, MassTransitMessagePublisher>();
-
-builder.Services.AddMassTransit(x =>
-{
-  var rabbitMqConfiguration = messagingConfig.Get<RabbitMqConfiguration>();
-  x.SetKebabCaseEndpointNameFormatter();
-
-  x.UsingRabbitMq((context, cfg) =>
-  {
-    var port = (ushort)rabbitMqConfiguration.Port;
-    cfg.Host(rabbitMqConfiguration.Hostname, port, rabbitMqConfiguration.VirtualHost, h =>
-    {
-      h.Username(rabbitMqConfiguration.UserName);
-      h.Password(rabbitMqConfiguration.Password);
-    });
-
-    cfg.ConfigureEndpoints(context);
-  });
-});
+builder.Services.AddMessaging(builder.Configuration);
 
 var assemblies = new Assembly[]
 {
@@ -96,23 +71,7 @@ builder.Services.AddInfrastructureDependencies(isDevelopment);
 
 var app = builder.Build();
 
-using (var scope = app.Services.CreateScope())
-{
-  var services = scope.ServiceProvider;
-  var hostEnvironment = services.GetService<IWebHostEnvironment>();
-  var loggerFactory = services.GetRequiredService<ILoggerFactory>();
-  var logger = loggerFactory.CreateLogger<Program>();
-  logger.LogInformation($"Starting in environment {hostEnvironment.EnvironmentName}");
-  try
-  {
-    var seedService = services.GetRequiredService<AppDbContextSeed>();
-    await seedService.SeedAsync(new OfficeSettings().TestDate);
-  }
-  catch (Exception ex)
-  {
-    logger.LogError(ex, "An error occurred seeding the DB.");
-  }
-}
+await app.SeedDatabaseAsync();
 
 app.UseResponseCompression();
